move state from Categories to App
